+++ jdaviz/configs/imviz/plugins/compass/compass.py
@@ -40,10 +40,6 @@ class Compass(PluginTemplateMixin, ViewerSelectMixin):
     def user_api(self):
         return PluginUserApi(self, expose=('viewer',), readonly=('data_label',))
 
-    def show(self, *args, **kwargs):
-        super().show(*args, **kwargs)
-        self._compass_with_new_viewer(from_show=True)
-
     def _on_viewer_data_changed(self, msg=None):
         if self.viewer_selected:
             viewer = self.viewer.selected_obj
@@ -61,7 +57,7 @@ class Compass(PluginTemplateMixin, ViewerSelectMixin):
         self.canvas_angle = viewer_item.get('canvas_angle', 0)  # noqa
         self.canvas_flip_horizontal = viewer_item.get('canvas_flip_horizontal', False)
 
-    @observe("viewer_selected", "plugin_opened_in_tray")
+    @observe("viewer_selected", "plugin_opened")
     def _compass_with_new_viewer(self, *args, **kwargs):
         if not hasattr(self, 'viewer'):
             # mixin object not yet initialized
@@ -69,7 +65,7 @@ class Compass(PluginTemplateMixin, ViewerSelectMixin):
 
         # There can be only one!
         for vid, viewer in self.app._viewer_store.items():
-            if vid == self.viewer.selected_id and (self.plugin_opened_in_tray or kwargs.get('from_show')):  # noqa
+            if vid == self.viewer.selected_id and self.plugin_opened:
                 viewer.compass = self
                 viewer.on_limits_change()  # Force redraw