+++ jdaviz/configs/imviz/tests/test_line_profile_xy.py
@@ -10,7 +10,7 @@ class TestLineProfileXY(BaseImviz_WCS_NoWCS):
     def test_plugin_linked_by_pixel(self):
         """Go through plugin logic but does not check plot contents."""
         lp_plugin = self.imviz.app.get_tray_item_from_name('imviz-line-profile-xy')
-        lp_plugin.plugin_active = True
+        lp_plugin.open_in_tray()
 
         lp_plugin._on_viewers_changed()  # Populate plugin menu items.
         assert lp_plugin.viewer_items == ['imviz-0']
@@ -66,7 +66,7 @@ class TestLineProfileXY(BaseImviz_WCS_NoWCS):
         assert lp_plugin.plot_available
 
         # Nothing should update on "l" when plugin closed.
-        lp_plugin.plugin_active = False
+        lp_plugin.plugin_opened = False
         self.viewer.on_mouse_or_key_event(
             {'event': 'keydown', 'key': 'l', 'domain': {'x': 5.1, 'y': 5}})
         lp_plugin.selected_x = '1.1'