+++ jdaviz/configs/specviz2d/plugins/spectral_extraction/spectral_extraction.py
@@ -95,7 +95,7 @@ class SpectralExtraction(PluginTemplateMixin):
     """
     dialog = Bool(False).tag(sync=True)
     template_file = __file__, "spectral_extraction.vue"
-    has_previews = Bool(True).tag(sync=True)
+    uses_active_status = Bool(True).tag(sync=True)
 
     active_step = Unicode().tag(sync=True)
     interactive_extract = Bool(True).tag(sync=True)
@@ -406,12 +406,12 @@ class SpectralExtraction(PluginTemplateMixin):
             else:
                 raise ValueError("step must be one of: trace, bg, ext")
 
-    @observe('show_previews', 'interactive_extract')
+    @observe('is_active', 'interactive_extract')
     def _update_plugin_marks(self, *args):
         if not self._do_marks:
             return
 
-        if not (self.show_previews):
+        if not (self.is_active):
             for step, mark in self.marks.items():
                 mark.visible = False
             return
@@ -469,22 +469,22 @@ class SpectralExtraction(PluginTemplateMixin):
 
         # the marks haven't been initialized yet, so initialize with empty
         # marks that will be populated once the first analysis is done.
-        self._marks = {'trace': PluginLine(viewer2d, visible=self.show_previews),
-                       'ext_lower': PluginLine(viewer2d, visible=self.show_previews),
-                       'ext_upper': PluginLine(viewer2d, visible=self.show_previews),
-                       'bg1_center': PluginLine(viewer2d, visible=self.show_previews,
+        self._marks = {'trace': PluginLine(viewer2d, visible=self.is_active),
+                       'ext_lower': PluginLine(viewer2d, visible=self.is_active),
+                       'ext_upper': PluginLine(viewer2d, visible=self.is_active),
+                       'bg1_center': PluginLine(viewer2d, visible=self.is_active,
                                                 line_style='dotted'),
-                       'bg1_lower': PluginLine(viewer2d, visible=self.show_previews),
-                       'bg1_upper': PluginLine(viewer2d, visible=self.show_previews),
-                       'bg2_center': PluginLine(viewer2d, visible=self.show_previews,
+                       'bg1_lower': PluginLine(viewer2d, visible=self.is_active),
+                       'bg1_upper': PluginLine(viewer2d, visible=self.is_active),
+                       'bg2_center': PluginLine(viewer2d, visible=self.is_active,
                                                 line_style='dotted'),
-                       'bg2_lower': PluginLine(viewer2d, visible=self.show_previews),
-                       'bg2_upper': PluginLine(viewer2d, visible=self.show_previews)}
+                       'bg2_lower': PluginLine(viewer2d, visible=self.is_active),
+                       'bg2_upper': PluginLine(viewer2d, visible=self.is_active)}
         # NOTE: += won't trigger the figure to notice new marks
         viewer2d.figure.marks = viewer2d.figure.marks + list(self._marks.values())
 
-        self._marks['extract'] = PluginLine(viewer1d, visible=self.show_previews)
-        self._marks['bg_spec'] = PluginLine(viewer1d, visible=self.show_previews, stroke_width=1)  # noqa
+        self._marks['extract'] = PluginLine(viewer1d, visible=self.is_active)
+        self._marks['bg_spec'] = PluginLine(viewer1d, visible=self.is_active, stroke_width=1)  # noqa
 
         # NOTE: += won't trigger the figure to notice new marks
         viewer1d.figure.marks = viewer1d.figure.marks + [self._marks['extract'],
@@ -497,7 +497,7 @@ class SpectralExtraction(PluginTemplateMixin):
              'trace_pixel', 'trace_peak_method_selected',
              'trace_do_binning', 'trace_bins', 'trace_window', 'active_step')
     def _interaction_in_trace_step(self, event={}):
-        if not self.show_previews or not self._do_marks:
+        if not self.is_active or not self._do_marks:
             return
         if event.get('name', '') == 'active_step' and event.get('new') != 'trace':
             return
@@ -518,7 +518,7 @@ class SpectralExtraction(PluginTemplateMixin):
              'bg_trace_selected', 'bg_trace_pixel',
              'bg_separation', 'bg_width', 'bg_statistic_selected', 'active_step')
     def _interaction_in_bg_step(self, event={}):
-        if not self.show_previews or not self._do_marks:
+        if not self.is_active or not self._do_marks:
             return
         if event.get('name', '') == 'active_step' and event.get('new') != 'bg':
             return
@@ -574,7 +574,7 @@ class SpectralExtraction(PluginTemplateMixin):
     @observe('ext_dataset_selected', 'ext_trace_selected',
              'ext_type_selected', 'ext_width', 'active_step')
     def _interaction_in_ext_step(self, event={}):
-        if not self.show_previews or not self._do_marks:
+        if not self.is_active or not self._do_marks:
             return
         if event.get('name', '') == 'active_step' and event.get('new') != 'ext':
             return