+++ jdaviz/configs/specviz2d/plugins/spectral_extraction/spectral_extraction.vue
@@ -2,9 +2,9 @@
   <j-tray-plugin
     description="2D to 1D spectral extraction."
     :link="'https://jdaviz.readthedocs.io/en/'+vdocs+'/'+config+'/plugins.html#spectral-extraction'"
-    :has_previews="has_previews"
+    :uses_active_status="uses_active_status"
     :plugin_ping.sync="plugin_ping"
-    :persistent_previews.sync="persistent_previews"
+    :keep_active.sync="keep_active"
     :popout_button="popout_button">
 
     <v-row>
@@ -28,8 +28,7 @@
     </v-row>
 
 
-    <div @mouseover="() => active_step='trace'"
-      :style="plugin_active && active_step==='trace' ? 'box-shadow: rgb(2 123 161 / 20%) -20px 0px 0px 0px' : ''">
+    <div @mouseover="() => active_step='trace'">
       <j-plugin-section-header>Trace</j-plugin-section-header>
       <v-row>
         <j-docs-link>
@@ -188,8 +187,7 @@
       </v-row>
     </div>
 
-    <div @mouseover="() => active_step='bg'"
-      :style="plugin_active && active_step==='bg' ? 'box-shadow: rgb(2 123 161 / 20%) -20px 0px 0px 0px' : ''">
+    <div @mouseover="() => active_step='bg'">
       <j-plugin-section-header>Background</j-plugin-section-header>
       <v-row>
         <j-docs-link>Create a background and/or background-subtracted image.</j-docs-link>
@@ -351,8 +349,7 @@
       </v-row>
     </div>
 
-    <div @mouseover="() => active_step='ext'"
-      :style="plugin_active && active_step==='ext' ? 'box-shadow: rgb(2 123 161 / 20%) -20px 0px 0px 0px' : ''">
+    <div @mouseover="() => active_step='ext'">
       <j-plugin-section-header>Extraction</j-plugin-section-header>
       <v-row>
         <j-docs-link>